move: 시간 관련 상수 constants 폴더로 이동
